+++ server/Repository/EmployeeRepository.js
@@ -1,0 +1,3 @@
+const UserModel = require('../Repository/UserRepository')
+const EmployeeSchema = require('../Schemas/Employee')
+module.exports =  UserModel.discriminator('Employee', EmployeeSchema)
